Make CI tests work
