+++ .github/workflows/release.yml
@@ -82,8 +82,25 @@ jobs:
 
       - name: Install dependencies
         run: |
+          sudo apt-get update
+          sudo apt-get install --no-install-recommends -y \
+            libxkbcommon-x11-0 \
+            x11-utils \
+            libyaml-dev \
+            libegl1-mesa \
+            libxcb-icccm4 \
+            libxcb-image0 \
+            libxcb-keysyms1 \
+            libxcb-randr0 \
+            libxcb-render-util0 \
+            libxcb-xinerama0 \
+            libopengl0 \
+            libxcb-cursor0 \
+            pulseaudio
+
           python -m pip install --upgrade pip
           python -m pip install -r requirements.txt
+          python -m pip install pytest-xvfb
 
       - name: Test with pytest
         run: |
